refactor: include filter controller in board controller. Disable filters when loading tickets
