+++ src/utils/utils.js
@@ -1,23 +1,13 @@
-export const SHOWING_TICKETS_ON_START = 5;
-export const SHOWING_TICKETS_ON_LOAD = 5;
-
-export const FilterTypes = {
-  all: false,
-  direct: false,
-  oneStop: false,
-  twoStops: false,
-  threeStops: false,
-};
-
-export const SortType = {
-  CHEAP: "cheap",
-  FAST: "fast",
-  OPTIMAL: "optmal",
+export const RenderPosition = {
+  AFTERBEGIN: `afterbegin`,
+  BEFOREEND: `beforeend`,
+  AFTEREND: `afterend`
 };
 
 export function convertMinuteToHM(minutes) {
-  let rhours; let minute; let
-    result = null;
+  let rhours;
+  let minute; 
+  let result = null;
   const hours = minutes / 60;
   rhours = Math.floor(hours);
   minute = Math.round((hours - rhours) * 60);
@@ -38,56 +28,3 @@ export function getArrivalTime(date, durationMinutes) {
   return `${h}:${m}`;
 }
 
-export const getCarrierLogo = (carrier) => `https://pics.avs.io/99/36/${carrier}.png`;
-
-export const getStopsSentense = (countStops) => {
-  switch (countStops) {
-    case 0:
-      return "без пересадок";
-    case 1:
-      return `${countStops} пересадка`;
-    case 2:
-      return `${countStops} пересадки`;
-    case 3:
-      return `${countStops} пересадки`;
-    default:
-      return `${countStops} пересадок`;
-  }
-};
-
-export function sortByPrice(ticketA, ticketB) {
-  if (ticketA.price > ticketB.price) return 1;
-  if (ticketA.price < ticketB.price) return -1;
-  return 0;
-}
-
-export const sortByFlightTime = (ticketA, ticketB) => {
-  const flightFrom = ticketA.segments[0].duration + ticketA.segments[1].duration;
-  const flightTo = ticketB.segments[0].duration + ticketB.segments[1].duration;
-
-  if (flightFrom > flightTo) return 1; if (flightFrom < flightTo) return -1;
-  return 0;
-};
-
-export const filterByStops = (tickets, stops) => tickets.filter((ticket) => {
-  const [first, second] = ticket.segments;
-  return first.stops.length === stops && second.stops.length === stops;
-});
-
-export const getSortedTickets = (tickets, sortType, from, to) => {
-  let sortedTickets = [];
-  const showingTickets = tickets.slice();
-
-  switch (sortType) {
-    case SortType.CHEAP:
-      sortedTickets = showingTickets.sort(sortByPrice);
-      break;
-    case SortType.FAST:
-      sortedTickets = showingTickets.sort(sortByFlightTime);
-      break;
-    default:
-      sortedTickets = showingTickets;
-  }
-
-  return sortedTickets.slice(from, to);
-};
